line_art/requirements.txt
numpy
pillow
matplotlib
IPython
